+++ src/components/_private/Flyout/Flyout.types.ts
@@ -1,6 +1,6 @@
 import React from "react";
 import type * as G from "types/global";
-import type { TrapMode } from "utilities/a11y/types";
+import type { TrapMode } from "utilities/a11y";
 
 /**
  * Utility