+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - Accordiong: fixed animation edge cases (inside live editor on the website)
 - useHotkeys: fixed mod+enter on non-mac os
 - Tabs: fixed edge cases when tab item active value is changing based on user events
+- useScrollLock: fixed ios lock
 - Icon: Fixed html width / height overrides
 - Tabs: Increased spacing between the items
 - Loader: Large size support
@@ -46,3 +47,5 @@
 - Flyout, Popover: initialFocusRef
 - View: removed additional div around the content when it has a single child
 - Flyout: Optimized resize observer to only be active when rendered
+- Overlay, Modal: Fixed scroll lock on safari
+- Fixed :focus-visible safari edge cases (back to data-rs-keyboard)